[#235] :recycle: refactor: 일기 등록 버튼 common 디렉토리 아래로 이동
