+++ client/src/components/common/diary/index.tsx
@@ -1,4 +1,5 @@
 import MoveableSticker from "./MoveableSticker";
 import DiarySection from "./DiarySection";
+import PostButton from "./PostButton";
 
-export { MoveableSticker, DiarySection };
+export { MoveableSticker, DiarySection, PostButton };
