--- client/src/components/diary/index.tsx
@@ -1,3 +1,0 @@
-import PostButton from "./PostButton";
-
-export default PostButton;
